FACE.py/requirements.txt
streamlit 
streamlit-webrtc 
opencv-python 
mediapipe 
numpy 
pillow 
av
